+++ Sources/Rules.swift
@@ -4083,7 +4083,7 @@ public struct _FormatRules {
                     wasDeclaration = false
                 }
                 let token = formatter.tokens[i]
-                outer: switch token {
+                switch token {
                 case .keyword("guard"):
                     isGuard = true
                 case .keyword("let"), .keyword("var"), .keyword("func"), .keyword("for"):
@@ -4103,14 +4103,8 @@ public struct _FormatRules {
                        formatter.currentScope(at: i) == .startOfScope("[")
                     {
                         if isDeclaration {
-                            switch formatter.next(.nonSpaceOrCommentOrLinebreak, after: i) {
-                            case .endOfScope(")")?, .operator("=", .infix)?,
-                                 .delimiter(",")? where !isConditional || formatter.currentScope(at: i) == .startOfScope("("):
-                                tempLocals.insert(name)
-                                break outer
-                            default:
-                                break
-                            }
+                            tempLocals.insert(name)
+                            break
                         }
                         argNames.remove(at: index)
                         associatedData.remove(at: index)