+++ Sources/Tokenizer.swift
@@ -2,7 +2,7 @@
 //  Tokenizer.swift
 //  SwiftFormat
 //
-//  Version 0.51.13
+//  Version 0.51.14
 //
 //  Created by Nick Lockwood on 11/08/2016.
 //  Copyright 2016 Nick Lockwood
@@ -551,10 +551,6 @@ extension UnicodeScalar {
             return false
         }
     }
-
-    var isSpaceOrLinebreak: Bool {
-        isSpace || "\n\r\u{000B}\u{000C}".unicodeScalars.contains(self)
-    }
 }
 
 // Workaround for horribly slow String.UnicodeScalarView.Subsequence perf
@@ -732,7 +728,7 @@ private extension UnicodeScalarView {
     }
 
     mutating func readToEndOfToken() -> String {
-        readCharacters { !$0.isSpaceOrLinebreak } ?? ""
+        readCharacters { !$0.isSpace && !"\n\r".unicodeScalars.contains($0) } ?? ""
     }
 }
 
@@ -742,18 +738,13 @@ private extension UnicodeScalarView {
     }
 
     mutating func parseLineBreak() -> Token? {
-        switch first {
-        case "\r":
-            removeFirst()
+        if read("\r") {
             if read("\n") {
                 return .linebreak("\r\n", 0)
             }
             return .linebreak("\r", 0)
-        case "\n", "\u{000B}", "\u{000C}":
-            return .linebreak(String(removeFirst()), 0)
-        default:
-            return nil
         }
+        return read("\n") ? .linebreak("\n", 0) : nil
     }
 
     mutating func parseDelimiter() -> Token? {
@@ -1476,10 +1467,8 @@ public func tokenize(_ source: String) -> [Token] {
             return
         }
         guard let prevNonSpaceIndex = index(of: .nonSpaceOrCommentOrLinebreak, before: i) else {
-            if string == "/" {
-                tokens[i] = .startOfScope("/")
-            } else if tokens.count > i + 1 {
-                tokens[i] = .operator(string, .prefix)
+            if tokens.count > i + 1 {
+                tokens[i] = string == "/" ? .startOfScope("/") : .operator(string, .prefix)
             }
             return
         }
@@ -1549,8 +1538,11 @@ public func tokenize(_ source: String) -> [Token] {
             guard let nextNonSpaceToken =
                 index(of: .nonSpaceOrCommentOrLinebreak, after: i).map({ tokens[$0] })
             else {
+                if prevToken.isLvalue {
+                    type = .postfix
+                    break
+                }
                 if token == .operator("/", .none),
-                   prevToken.isSpaceOrLinebreak ||
                    prevNonSpaceToken.isOperator(ofType: .infix) || (
                        prevNonSpaceToken.isUnwrapOperator &&
                            prevNonSpaceIndex > 0 &&
@@ -1562,9 +1554,6 @@ public func tokenize(_ source: String) -> [Token] {
                    ].contains(prevNonSpaceToken)
                 {
                     tokens[i] = .startOfScope("/")
-                } else if prevToken.isLvalue {
-                    type = .postfix
-                    break
                 }
                 return
             }
@@ -1875,12 +1864,6 @@ public func tokenize(_ source: String) -> [Token] {
         token = tokens[count - 1]
         switch token {
         case .startOfScope("/"):
-            if let next = characters.first, next.isSpaceOrLinebreak {
-                // Misidentified as regex
-                token = .operator("/", .none)
-                tokens[count - 1] = token
-                return
-            }
             scopeIndexStack.append(count - 1)
             let start = characters
             processStringBody(regex: true, hashCount: 0)